@@ -1,11 +1,9 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 
-import App from './App.vue'
-import store from './store'
+import App from '@/App.vue'
+import store from '@/store'
 
-Vue.config.productionTip = false
-
-export default new Vue({
-  store,
-  render: h => h(App),
-}).$mount('#app')
+// @ts-ignore
+const app = createApp(App)
+app.use(store)
+app.mount('#app-root')
