@@ -1,6 +1,6 @@
 import TileValueUnit from '@/enums/tileValueUnit'
 
-export default interface TileValue {
+export default interface TileMetrics {
   values: string[],
   unit: TileValueUnit,
 }
