@@ -1,13 +1,13 @@
 import TileStatus from '@/enums/tileStatus'
 import TileType from '@/enums/tileType'
 import TileBuild from '@/interfaces/tileBuild'
-import TileValue from '@/interfaces/tileValue'
+import TileMetrics from '@/interfaces/tileMetrics'
 
 export default interface TileState {
   type: TileType,
   status: TileStatus,
   label?: string,
   message?: string,
-  value?: TileValue,
+  metrics?: TileMetrics,
   build?: TileBuild
 }
